feat:4.21

+++ controllers/blogs.js
@@ -20,7 +20,6 @@ blogsRouter.get("/:id", async (request, response) => {
 })
 
 blogsRouter.post("/", async (request, response) => {
-  console.log(request.token)
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error:"invalid token" })
@@ -53,6 +52,16 @@ blogsRouter.put("/:id", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async(request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error:"invalid token" })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const blogCreatorId = blog.user.toString()
+  if (blogCreatorId !== decodedToken.id) {
+    return response.status(401).json({ error: "You dont have permission to delete this blog" })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
